bridge-bot/requirements.txt
config==0.5.1
motor==2.5.1
discord.py==2.4.0
checksumdir==1.2.0
